Move Baseline generator to its own solution
