+++ tools/MUXBaselineResourcesGenerator/MUXBaselineResourcesGenerator/App.xaml
@@ -1,7 +1,7 @@
 ﻿<Application
-    x:Class="BaselineResourcesGenerator.App"
+    x:Class="MUXBaselineResourcesGenerator.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:BaselineResourcesGenerator">
+    xmlns:local="using:MUXBaselineResourcesGenerator">
 
 </Application>
