+++ tools/MUXBaselineResourcesGenerator/MUXBaselineResourcesGenerator/Assets/SplashScreen.scale-200.png
