+++ tools/MUXBaselineResourcesGenerator/MUXBaselineResourcesGenerator/Assets/Square44x44Logo.scale-200.png
