+++ tools/MUXBaselineResourcesGenerator/MUXBaselineResourcesGenerator/Assets/StoreLogo.png
