+++ tools/MUXBaselineResourcesGenerator/MUXBaselineResourcesGenerator/MainPage.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="BaselineResourcesGenerator.MainPage"
+    x:Class="MUXBaselineResourcesGenerator.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:BaselineResourcesGenerator"
+    xmlns:local="using:MUXBaselineResourcesGenerator"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
@@ -26,7 +26,9 @@
             <Button Content="Copy" Name="Copy" Click="Copy_Click" Grid.Column="1" />
         </Grid>
 
-        <TextBlock Name="BaselineTextBox" Grid.Row="1" />
+        <ScrollViewer VerticalScrollMode="Enabled" HorizontalScrollMode="Auto" Grid.Row="1">
+            <TextBlock Name="BaselineTextBox" />
+        </ScrollViewer>
 
     </Grid>
 </Page>
