+++ tools/MUXBaselineResourcesGenerator/MUXBaselineResourcesGenerator/Package.appxmanifest
@@ -7,14 +7,14 @@
   IgnorableNamespaces="uap mp">
 
   <Identity
-    Name="0e7026b0-ff70-401e-b089-41ed8b718072"
+    Name="abace1ca-ea24-4e89-ad78-2c0db275a83e"
     Publisher="CN=almedina"
     Version="1.0.0.0" />
 
-  <mp:PhoneIdentity PhoneProductId="0e7026b0-ff70-401e-b089-41ed8b718072" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
+  <mp:PhoneIdentity PhoneProductId="abace1ca-ea24-4e89-ad78-2c0db275a83e" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
 
   <Properties>
-    <DisplayName>BaselineResourcesGenerator</DisplayName>
+    <DisplayName>MUXBaselineResourcesGenerator</DisplayName>
     <PublisherDisplayName>almedina</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -30,12 +30,12 @@
   <Applications>
     <Application Id="App"
       Executable="$targetnametoken$.exe"
-      EntryPoint="BaselineResourcesGenerator.App">
+      EntryPoint="MUXBaselineResourcesGenerator.App">
       <uap:VisualElements
-        DisplayName="BaselineResourcesGenerator"
+        DisplayName="MUXBaselineResourcesGenerator"
         Square150x150Logo="Assets\Square150x150Logo.png"
         Square44x44Logo="Assets\Square44x44Logo.png"
-        Description="BaselineResourcesGenerator"
+        Description="MUXBaselineResourcesGenerator"
         BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png"/>
         <uap:SplashScreen Image="Assets\SplashScreen.png" />
