+++ tools/MUXBaselineResourcesGenerator/MUXBaselineResourcesGenerator/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("BaselineResourcesGenerator")]
+[assembly: AssemblyTitle("MUXBaselineResourcesGenerator")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("BaselineResourcesGenerator")]
+[assembly: AssemblyProduct("MUXBaselineResourcesGenerator")]
 [assembly: AssemblyCopyright("Copyright ©  2021")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
