+++ tools/MUXBaselineResourcesGenerator/MUXBaselineResourcesGenerator/Properties/Default.rd.xml
